- refactor: decomposed monolithoc main.c into dedicated files, following similar structure as what is find in `blockBlast` branch
