--- lobby/include/common.h
@@ -1,11 +1,0 @@
-#ifndef COMMON_H
-#define COMMON_H
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include "raylib.h"
-#include "raymath.h"
-
-#endif
