--- lobby/include/config.h
@@ -1,7 +1,0 @@
-#ifndef CONFIG_H
-#define CONFIG_H
-
-#define WINDOW_WIDTH 800
-#define WINDOW_HEIGHT 600
-
-#endif
