--- lobby/include/constant.h
@@ -1,15 +1,0 @@
-#ifndef CONSTANT_H
-#define CONSTANT_H
-
-#define GRAVITY 1200.0f
-#define MOVE_SPEED 300.0f
-#define JUMP_FORCE 500.0f
-#define GROUND_Y 50
-
-#define COYOTE_TIME 0.1f
-#define JUMP_BUFFER_TIME 0.1f
-#define MAX_JUMPS 2
-
-#define FRICTION 2000
-
-#endif
