--- lobby/include/global.h
@@ -1,20 +1,0 @@
-#ifndef GLOBAL_H
-#define GLOBAL_H
-
-#include "types.h"
-
-extern Platform_st platforms[];
-extern int platformCount;
-
-extern Texture2D playerTextures[];
-extern int playerTextureCount;
-
-extern Rectangle defaultPlayerTextureRect;
-
-extern Rectangle skinButtonRect;
-
-extern bool isTextureMenuOpen;
-
-extern Texture2D logoSkinButton;
-
-#endif
