+++ lobby/include/ui/app.h
@@ -1,0 +1,7 @@
+#ifndef UI_APP_H
+#define UI_APP_H
+
+void drawMenuTextures(void);
+void drawSkinButton(void);
+
+#endif // UI_APP_H
